Experimental addition of locking to `ContentProvider` with a custom `Lock`
There are a few things added here, notably `parking_lot` as a dependency,
but the most significant is the addition of a new lock, defined in
'views/file/lock.rs'. Documentation has not been written yet, but the
general overview is as follows:

The lock has been designed to allow simultaneous immutable borrows *and*
global lock/unlock so that immutability can be guaranteed between borrows.
The alternative was to allow some sort of upgradable read lock as part of
the `ContentProvider` definition, and that seemed like it would be too
much hassle to deal with in the simple case of implementing for `Lines`.