Initial changes to parsing key inputs in Normal mode
This first version is a rough draft - it dramatically reduces the
features that are actually available, but provides the infrastructure
to create many more.

The overarching change is that normal mode now allows custom parsing of
key inputs in a way that will allow them to be joined together
(e.g. number + 'd' + number + movement for delete).