Refine multi-line delete
This commit changes the behavior of multi-line delete (i.e. key inputs
of the form "d{n}(j|k)") to align more closely with what would be
expected of a Vi derivative. Where previously the behavior was to
delete the region between the two positions, now all of the lines moved
across are deleted.

This required adding a new method - `ViewBuffer::delete_lines` - and making
public `ViewBuffer::simulate_movement`.