Remove freedom for potential deadlocks from `ContentProvider::lock`
This commit changes `ContentProvider::{lock,unlock}` to require mutable
access to the type being locked. This is because a potential source of
deadlocks could occur from locking while the current thread has a read
lock.

The documentation for `views::file::lock::Lock` has been changed
accordingly to indicate proper use; the motiviation for this change is
described there in detail.