Moves locks to the crate level, adds a deadlock-safe wrapper: `ArcLock`
The original `Lock` type has also been renamed to `RawLock` to
discourage use and highlight the risky nature of its API. Details on the
restrictions that allow `ArcLock` to be safe are described in the
"deadlock safety" section of the `RawLock` documentation. The relevance
here is that the new restrictions required a few methods to be changed
where they would have previously allowed deadlocks.

Additionally, a small bug was fixed: `views::file::handle::File`
previously failed to revert `unsaved` to false after writing a file to
the filesystem. This has now been changed.