Refactor: Move `normal::Parsers` into 'combinators.rs', rename to `Set`
This commit makes a change with no difference in functionality. The
bulk of the parsing work in 'normal/mod.rs' was being done by a type
that could have been abstracted over, so it is now. The type was
previously `Parsers`, and represented a static set of parsers that were
used for normal mode, but it has been renamed to `Set` and produces
output of any kind given any list.

A new enum, `SetResult` has been added as a wrapper around the output
of `Set` to handle the different possible failure modes (this was not
done previously). More information can be found in the type's
documentation.