Improve 'de' keybinding in normal mode
Background:
Because movement-based deletion excludes the final character in the
range, deleting to the end of a word did not behave in the manner that
would typically be expected - i.e. deleting *through* the end of the
word. Instead the cursor would stop at the end of the word and the
deleted range would (like everything else) exclude the cursor's current
position.

Solution:
This commit solves the problem by changing a certain subset of movements
before the deletion is acted on; rightwards character predicates that
"shift back" (`UntilFst`) are converted to not shift (`UntilSnd`).

Looking forward:
This commit may have unintended side-effects in the future because of
the nature of the change. A better, more permanent solution would be to
instead add a separate parameter to deletion in order to allow for
specific choice in what type of range to treat the movement as
(i.e. inclusive vs. exclusive).