Remove `resize` method from `views::View`
The `resize` method created multiple sources of truth because each `Painter` was also giving a size; individual `View`s relied on the painter's size being identical to the last resize they were given. This has now been changed so that the responsibility for resizing is also handled in `View::refresh` whenever the painter's size changes.