Complete implementation of mode configuration
This commit provides a working implementation of the intended
configuration extension available for modes. This is demonstrated by the
addition of normal mode keybindings for changing which view is focused.

Also included in the set of changes here is the removal of both prelude
modules that were previously present. The intention is to re-add the
`config` prelude as it originally was, once the necessary items become
clear.

This branch is not yet ready to merge; there is still documentation in
the `mode` module that must be written/revised.