Fix width-out-of-bounds crash on undo
There was a bug in the way that diffs were handled in
`ViewBuffer::refresh_diffs` - the diffs are applied before being handed
ot the function, but updating the cursor's byte position was done with
the old row/column. This meant that the following set of actions could
consistently cause a panic from the column being out-of-bounds:

 1. Append 'foobar' to the end of a line
 2. With the cursor within 'foobar', undo the change

This caused a panic because the cursor's width was suddenly past the
end of the line. The solution here was to separately track the cursor's
byte index within the file, using the addition of a field
`cursor_byte_idx` to the `ViewBuffer` struct.