Fix panic on `d$`
This was due to `simulate_movement` returning `usize::MAX` for a column
index when moving to the end of the line, which was subsequently used in
`delete_movement` directly to obtain a byte index.

The change simply bounds the upper value of the deletion range by the
width of the line.