Fix split bugs (x2)
The first bug was due to a failure to resize after adding to a split -
this meant that splits would erroneously indicate they could not display
when it was actually possible. This was fixed by the small additions of
calls to `resize()` in the handling of `Close` and `Open` output
signals.

The second was an issue with shifting focus across vertical splits, e.g.
when the split layout looked like:
  +--+--+-----+
  |  |  |     |
  +--+--+     |
  |     |     |
  +-----+-----+
and we try to move into the large split on the right from the top-right
split on the left side. This was simply due a typo in the handling of
the `ShiftFocus` output signal in the `Vert` split.