Fix selected uppercase keybindings
Due to a recent change in crossterm, uppercase characters in key events
always gave `Shift` - instead of the previous implementation, which only
gave shift on Windows. This happened to break a few keybindings, which
wasn't noticed until later.

This has now been fixed with the introduction of a helper function for
`KeyEvent`: `shift`, which produces a `Char` key event with the shift
key pressed. Determining which keybindings needed changing was simply
done with ripgrep:

  rg -p "::none\\('[A-Z]'"