Fix panic when file does not exist, add `crate::fs`
Previously, the editor would panic when called with a path that did not
point to a preexisting file. Part of the fix has been to create a custom
`fs` module at the crate root that guarantees all constructed paths have
been canonicalized.

Every other user of `std::fs` in the source has been moved to
`crate::fs` and `crate::fs::Path`.