improve `#[derive(Typed)]` type bounds
Originally, we used a "where" clause for every single mentioned type.
That's *technically* correct, but produces a lot of warnings, because
for concrete types the answer of "does this type implement the trait"
has a direct answer (instead of, say: "only if T implements XY").

So we switched to the naive sort of binding that most derive macros use:
simple requirements on any generic type parameters. But this doesn't
work for things like `Vec<T>`, which requires `T: Clone` to implement
`Typed`.

This implementation notices any types that contain a referenced generic
type parameter and includes those in the bounds; with this method, we
should get the best of both worlds.