Fix warnings from 'cargo doc'
There's still a few left from references to `Cursor` in src/text/mod.rs,
but those seem to be erroneous. Everything else was fixed.

There was also a place in 'history_core' that required adding a method
that had been referenced but never written -- it's a pretty simple
method, so I've just included it in this commit instead of a separate
thing.