Improve `EditId` recycling scheme
Previously, `HistoryCore` stored a list of `Option<Edit>`, and searched
through the list to find an empty slot to put a new one. This commit
changes it to use an embedded linked list between indexes to track where
all of the empty slots are.

This changes it from O(n) to O(1) on both removing *and* adding an edit.