Provide stronger guarantees for `Ranged` slices
This change essentially comes to clean up the implementation of `Ranged`
a bit. Previously, we'd always store the root node as `Some(_)`, even if
that didn't really make sense. This meant we sometimes had to do silly
things, like slitting a slice at the very start or end, which produces
slices of length zero.

We don't want to allow lengths equal to zero *anwyays*, so this change
does two things:

  1. Ensures that `self.root` is `None` iff `self.size` is zero, for any
     `Ranged`.
  2. Guarantees that calls to `AccumulatorSlice::split_at` will never
     ask for a slice with length zero -- i.e. the index will never be
     equal to zero or the length of the slice.

The test code was changed slightly in order to check the second of these
conditions.

We also added `[Std]Ranged::new_empty` as a new way to construct
`Ranged` trees without providing an initial value, now that we don't
actually need to.