Provide more guarantees for `Ranged` semantics
This commit adds a couple more guarantees about how the accumulators are
used in a `Ranged`. The full description is in the documentation for
`AccumulatorSlice::Accumulator`, but the gist is that operations will be
minimal: addition never overlaps and subtraction is always corresponds
to a sub-range, sharing either the lower or upper edge.

There's a significant amount of changes to the testing code in order to
allow checking that these are actually upheld.