Add stable node references to `Ranged`
Yet another change in the interest of implementing a certain data
structure. This one adds stable node references to `Ranged`, so that we
can go from Node -> position in O(log n) as well.

This... made a lot of the code messier. I'm not sure what the best way
to resolve that is. There's a few things planned, like a custom pointer
type similar to `Rc<RefCell<T>>`, but on the whole we can't really get
around the problem of needing to use guards everywhere to access the
contents of nodes.

Thankfully, this complexity shouldn't add *too* much actual overhead,
both because single-threaded reference counting tends to be very fast,
and because our accessing within the data structure still largely
follows a single-ownership pattern.

---

The primary export from this change is the new `NodeRef` type, defined
in 'src/text/ranged/core.rs'. An abundance of information about it is
given with its definition.