Fix `NodeRef::collapse_redirects`
In short, there were many problems with this method as written. There
were many cases where it resulted in panics from borrowing in the
destructor of a `NodeRef` we drop along the way, in addition to a
multitude of other problems.

I'm planning on implementing a custom pointer type so that we don't need
to deal with this in the first place, but it's best to have a working
solution in the meantime.