Guarantee `NodeRef`s move left on splits, +fix
The primary change here is to guarantee that, when a `Node` splits, any
references to it stay with the left-hand side of that node. This gives a
base of deterministic behavior to build other things on top of.

In the course of making this change, I found a bug in the implementation
of `tests::Slice` in that `accumulated_at` didn't behave properly when
not called with full length of the slice. This is also fixed here.