Implement `RelativeSet` (initial commit)
We add the `RelativeSet` type here, which will be useful in the near
future for tracking the positions of items where groups must be shifted
(e.g. line numbers, shifted after a newline is inserted).

The only notable changes outside of the new file, '.../ranged/set.rs'
are:
 * Certain values/methods in ranged/core.rs are now `pub(super)`, in
   order to give `RelativeSet` access to them in its implementation.
 * `ranged::RangedIndex` has been given the `increment` method so that
   we can construct ranges of the form `idx..idx + 1`.

For more information about `RelativeSet`, have a look at the docs!

---

As an aside: Currently, we can't get rustdoc to test the examples in the
methods. To get around this, each documented item with an example is
given its own test function in the `doctests` module at the bottom of
the file. Each example is also tagged with its own `@def` to track
changes, just to help ensure that the examples stay consistent with
what's actually being tested.