Allow `ranged::rc::Weak` to dangle for fn `new`
The previous version of `Weak` always pointed to a valid allocation.
This has some nice properties, but it also means that constructing an
always-invalid reference has an associated cost it doesn't need.

So we took a little trick from the standard library to create dangling
pointers we always know are dangling. There's a bit of explanation in
the new comment inside `Weak`.