Add new methods to `Borrow`/`BorrowMut`
There's actually a few interesting things going on here. In order to
support borrow mapping, we had to wildly change the types used for
borrows. The correctness of that change is... not entirely simple, and
so there's a hefty comment above `Borrow` explaining my reasoning for
why I believe the API is sound.

This commit also adds a dependency on #![feature(dropck_eyepatch)] in
order to provide #[may_dangle] on the relevant types.