Add `BorrowMut::stack` for accessing deep values
BACKGROUND:

In order for `OwnedCell` to be used in the implementation of `Ranged`,
we need to support more of the borrowing semantics that were originally
guaranteed by accessing behind mutable references.

When using `Box` as the smart pointer for nodes, we could safely extract
singular mutable borrows deep inside the tree, just by following the
pointers. We still want to be able to do that in the implementation of
`splay`, and the existing solution of cloning the reference is no longer
available with `OwnedCell`.

SOLUTION:

This commit adds the `stack` method to `BorrowMut`, which enables
extracting borrows behind multiple `OwnedCell`s. It's not perfect,
because we still need to mark every node as borrowed (and then
un-borrowed), but it's a sound solution that works within the
restrictions of the `OwnedCell` API.