Fix double-free: add `State::DropInProgress`
There was a bug to do with dropping self-referential data structures
where dropping an `OwnedCell` that contains the last `Weak` pointer to
itself would free the associated memory twice.

So now there's a copule tests related to that. The only real solution
here was to add a new state to mark a value as "currently being
dropped", so that's what's been implemented.